Fix for bug #36009: Rendered ortho view messes up Freestyle lines in 3D viewport.
Clipping start is negative when the viewport preview is used with the orthographic view,
while Freestyle assumes that imported mesh data are in the camera coordinate system
with the view point located at origin.  The present solution is to adjust the clipping
start/end and introduce a Z-axis offset for mesh vertices so as to satisfy the assumption.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@58129 954f8c5b-7b00-dc11-b283-0030488c597c