Dynamic Paint: Added a new "smoothness" parameter for waves.
It greatly helps getting rid of that "noise" that occurs if you use really steep objects (like cubes) as a brush. New default value is 1.0 which is just high enough to only get rid of the sharpest spikes, so if you want really smooth waves it's better use higher values.

This also seems to "fix" bug [#35413].

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@58850 954f8c5b-7b00-dc11-b283-0030488c597c